+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- CLI: fix the display of the assets in yaml and json format (#115)
 - No more crash when running a script in local subprocess mode from a path with spaces (#99)
 
 ## [0.14.0] - 2021-12-01